Move NettyByteBody
