+++ http-netty/src/main/java/io/micronaut/http/netty/body/BodySizeLimits.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.http.server.netty.body;
+package io.micronaut.http.netty.body;
 
 import io.micronaut.core.annotation.Internal;
